Connection Changes
Added link to source of SerialTest code.
Minor formatting changes.
initialize() in SerialTest now returns a bool indicating whether or not
the connection to the serial port was successful. This is used in the
GUI for determining whether or not to update the Connect/Disconnect
buttons accordingly. An error appears in the text pane if the connection
fails.