Arduino Receiving Input
Added an input field for the Arduino. Modified the Arduino code to read
input: terminates on '/' that is appended by the sendButton. This causes
an issue if there is a '/' in the string being sent, so currently the
Java side of things removes the slashes first.
The main frame is no longer resizable.
disableElements() now takes care of disabling form elements during a
disconnect/failed connection.
A prefix (SENT:, REC:, ERROR:) should now be required when writing
anything to the serialOutputTextPane.